refactor: simplify delight-config follow-on per /simplify pass
Cleanups from the three-agent /simplify review (reuse, quality,
efficiency):

- Extract shared `personality::format::format_bytes` (was duplicated
  in narration.rs and summary.rs).
- Add `MediaType::is_photo_like` / `is_video_like` to consolidate the
  match arms duplicated across the streaming and cleanup success
  paths.
- Add `DownloadTask::to_recap_asset` so each success site is a one-line
  recap observation instead of a 10-line struct literal.
- Gate `recap.observe(...)` on `mode.is_friendly()` so off-mode skips
  the per-asset filename String allocation.
- Drop the pre-cycle `db.get_summary()` query; derive `library_before`
  by saturating-subtracting the cycle's bytes from the post-cycle
  total. One SQL aggregate per cycle instead of two.
- `RunRecap::observe` switches the album map to `Entry::Occupied` /
  `Vacant` so the present-key path doesn't clone-then-drop the asset
  via `or_insert(asset)`.

Net -79 lines, no behaviour change. 2487 lib tests still pass; fmt
and clippy clean.
