+++ src/state/types.rs
@@ -137,6 +137,19 @@ impl MediaType {
             _ => None,
         }
     }
+
+    /// True for `Photo` and the still-image side of a Live Photo. Used by
+    /// the friendly summary card to bucket the cycle's downloads as
+    /// "photos" vs "videos" without each call site re-spelling the match.
+    pub fn is_photo_like(self) -> bool {
+        matches!(self, Self::Photo | Self::LivePhotoImage)
+    }
+
+    /// True for `Video` and the moving side of a Live Photo. Mirrors
+    /// `is_photo_like`; the two are exhaustive over `MediaType`.
+    pub fn is_video_like(self) -> bool {
+        matches!(self, Self::Video | Self::LivePhotoVideo)
+    }
 }
 
 /// Provider-agnostic metadata for an asset.
